Update following latest review
